Create LineToCross component
